requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.7
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
urllib3==2.1.0
